Added land enemies that move with the background

+++ Level/Level.h
@@ -45,7 +45,7 @@ private:
 	sf::Texture backgroundImg;
 	sf::Texture playerImg;
 	sf::IntRect rect;
-	int backgroundSpeed = 1, backgroundDist = 0;
+	float backgroundSpeed = 1, backgroundDist = 0;
 	short country;
 
 	// Maybe put it in game or some other texture storage idk
